+++ lilya/conf/enums.py
@@ -1,6 +1,21 @@
 from enum import Enum
 
-from dymmond_settings.enums import EnvironmentType as EnvironmentType
+
+class BaseEnum(str, Enum):
+    def __str__(self) -> str:
+        return self.value  # type: ignore
+
+    def __repr__(self) -> str:
+        return str(self)
+
+
+class EnvironmentType(BaseEnum):
+    """
+    An Enum for environments."""
+
+    DEVELOPMENT = "development"
+    TESTING = "testing"
+    PRODUCTION = "production"
 
 
 class StrEnum(str, Enum):